Added UserStatus RPC
Added the UserStatus RPC which replaces the old Login RPC. This new
functionality allows for a client to connect to the server and is is
then registered as online, the server keeps a record of connected
clients. This functionality will facilitate further addition of  Client
to Client chat creation, but only in real time for the moment.

- Added UserStatus RPC and supporting message types to message.proto
- Added UserStatus() to server, which registers clients that connect to
  the server.
- Modified Client's Login() to make use of the new connection stream by
  connecting users to a UserStatus stream as a form of Online/Offline
  status.
- Modified preparedstatements.go and init.sql to reference username,
  rather than uname, we can afford to have more verbose variables.
- Updated Makefile to remove unnecessary directory aliases to stop it
  interfering with make autocomplete.