Added Intial BeginChat() functionality
Added the inital ability to notify another user that your client wants
to connect and start a chat. As there is no prompt for user input yet,
the client will connect to streams for status and message and those
streams will be recorded on the server. The client then calls the
BeginChat() RPC and provides the username of the client it wants to chat
with. The RPC checks for connected users, and if they are online, it
then collects its message stream and sends a system notification that a
client wishes to chat with it.

- Added the BeginChat() server implementation that notifies clients if a
  user wants to chat with them.
- Added the cleint side implementation to take a username and have that
  user notified about a potential chat.
- Updated protobuf definition for GetMessages to take a username as
  input, which then registers the clients message stream with the
  server.

Current implementation is hacky and not thread safe, and contains hard
coded values but spinning up a second client container with a
client12345 user provided in environment variables yeilds a successful
BeginChat prompt.