Added Input loop for Strike Client
Now that we can create and confrim chats, we need to implement an input
loop rather than hard coding and chaning requests for creation and
confirmation. This added input functionality will allow for commands,
such as the implemented /login command, to be executed by the user while
still maintaining control of their client.

- Added input loop to the main client loop, it accepts /login and /exit
  commands.