Refactored Get/SendMessages
Refactored the GetMessage and SendMessage RPC implemetations to make use
of channels rather than interacting with streams directly. Clients will
now call GetMessages, and register their stream and channel, in a
goroutine, Messages will be sent to a client as they are added to the
respective channel. SendMessages now checks the recipient of an envelope
and queries the channel map for that user, if found, it will push the
envelope to their channel.

- Updated GetMessages and SendMessages to make use of channels rather
  than direct streams.
- Updated the Envelope type for the time being in the protobuf
  definitions to contain sender and receiver.
- Added the ability to supply arguments to commands in the interactive
  user shell, i.e. /beginchat username