Implemented Client to Client comms
Added the ability for clients to use the Updated GetMessages and
SendMessages RPCs. Clients now have access to an interactive shell for
messaging, /msgshell, which initiates their message stream, and allows
them to review messages prefixxed with their username i.e.
client0:hello. This functionality is still pretty rough, but 2 client
containers are able to send and recieve messages.

- Added Map initialization to server for the channel map used for
  sending client messages.
- Updated ConnectMessageStream() to print messages, and handle the
  potential for Stream EOF.
- Added SendMessage() to the client to correctly make use of the
  SendMessages() RPC.
- Added MessagingShell() to provide an interactive Chat shell for
  clients.
- Added a command /msgshell to launch the Chat Shell from the main
  cleint loop.