Refactored Client key loading
Refactored how the client loads its encryption and signing keys to be
far less verbose. Defined enums in the keys package for both Signing and
Encryption key types, with an added KeyDefinition structure, allows for
a map of all keys to be created and passed to the new
LoadAndValidateKeys() function found in keys.go. This new function takes
the path and makes use of GetKeyFromPath and the respective Validation
functions to handle key extraction from Pem files, and returns a map of
loaded keys ready to be used. This greatly reduces the amount of code
present in the Main client function, considering 4 keys need to be
loaded and validated.

- Added LoadAndValidateKeys() to keys.go to simplify parsing user key
  files.
- Removed Verbose key loading process found in the Client main function.
- Removed the now defunct AutoChat() function