Rectified Issue with Password verfification
Made a couple of modifications to the Password package, In HashPassword,
we now pass in our salt as a variable, removing the randomness from that
function itself. Doing this means that it can be used to hash passwords
client side before they go over the wire to the server. This is
supported by a new RPC SaltMine to get the relevant users salt that was
used for hashing. Removed most of the VerifyPassword functionality, as
it was wrongly hashing an already hashed password.

- Updated HashPassword to take salt as a parameter. Salt is now
  generated in the Signup, and retrieved in the login, where a hashed
  password is then sent for verification server side.
- Updated Init.sql to include column for user salts.
- Added relevant prepared statement for salt retrieval.
- Updated dockerfile for client to reduce build time and over all image
  size.
- Added a new RPC for getting user Salt.