Schema/Protobuf definitions for chat implemetation
Added Schema definitions for chats and basic message types in init.sql,
this should be a good starting point to create chats and persist
messages between sessions.

Added Protobuf definition for new stream MessageStreamPayload which can
work with a variety of types, allowing for a singular stream to be
created on login, and the user capable of creating/confirming chats, and
them messaging with those chats.

- Schema definitions for chats and messages
- Protobuf definition for new advanced message stream