Began implementing new Message Stream payload type
Modified the SendMessage function to support the new
MessageStreamPayload which can contain begin/accept chat, or an envelope
message. Similar functionality to how messages previously worked, but
now the same users message stream can have a variety of types provided
to it, which should allow for a better flow handling for chat creation
and use.

- Updated Server's SendMessages() and BeginChat() functions to support
  the new payload type
- Added MessageStream() to the server to replace GetMessages
  functionality.
- Updated payload type to container target username.