Replaced Client GetMessages with MessageStream
Implemented client side useage of MessageStream to allow for Envelope{}
and BeginChatRequest{} types to be sent over the same wire. Modified the
output types of both SendMessages and BeginChat to make use of the
MessageStreamPayload{} type that contains both chat request/responses
and envelopes. Added a type switch to connect MessageStream, which
checks the payload type, and sends the relevant output the the clients
shell, be that a message or a chat request that was recieved.

- Implemented client side changes in order to make use of the
  MessageStream for multiple message types.
- Removed BeginChatRequest RPC from the protobuf definition as it is now
  part of MessageStream.