Added ability to name chats
Added the functionality to name a Chat when creating a request, another
step in making the messaging shell more configurable, allowing a user to
switch between their chats.

Added a State enum to Chats in the protobuf definition, this should help
with building towards enabling a Diffie-Hellman key exchange. Chats will
be created with an INIT state, when 2 clients complete the request ->
invite acceptance flow, we should trigger a key exchange here and move
that INIT state to Active.

- Made chat naming configurable
- Added State enum to Chat types for future key exchange functionality
- Removed some comments