Finished refactoring MessagingShell
Moved the default case Send Message handling to its own function,
shellSendMessage, allowing for removing the ugly switch case for
handling commands in the MessagingShell function.

Commands are now part of a map, so input that matches the commands name
will be referenced against that map and fire the relevant shell*
command.

Refactored most of the shell* functions to a be a bit more clean and
concise.

- Finished refactoring MessagingShell to remove the switch statement.
- Modified existing shell* functions to be a bit cleaner
- Added function to print /help