Added KeyExchangeRequest Demultiplexer processing
Added functionality to the Demultiplexer to process and automatically
respond to a key exchange request that is recieved. This required some
modifications allowing for the Demultiplexer to access the user keys
that have been loaded into config.

ProcessKeyExchangeRequests will derive a shared secret using the
ComputerShareSecret function and the key exchange request, and then use
ReciprocateKeyExchange to finish out the key exchange.

- Added ProcessKeyExchangeRequests to Demultiplexer
- Modifications made to allow the Demultiplexer the use of users keys
- Added key exchange message payloads to protobuf definitions