Added key derivation function
Added DeriveKeys which takes the computed shared secret of a key
exchange, and generates the symetric Encryption and HMAC keys that will
be used when encrypting messages.

Shared secret computation can be carried out efficiently, persistence is
a potential vulnerability, so the shared secret and it's derived keys,
should be cached when chat is set to active.