Begun modifying login behaviour
Removed the status stream creation from the login function, moving it to
RegisterStatus(). Removing the online status from the login function
will allow for an easier time modifying how the client connects to a
server.

Changed the LoginUser prepared statment on the server side to rely on
usernames rather than passwords, the previous behaviour where we would
retireve the userID duing login was poorly implemented.