Updated existing message passing functions
Updated currently implemented Message sending code to make use of the
updated cache.

Cache now holds a more detailed set of data related to the current chat.
Rather than relying on an active chat ID and then traversing a cached
list of chats, all data related to the current chat is in a single cache
object.