Modifed payload processing workers
In order to preserve a clean and correct approach to concurrency and
context, the payload processing scaling needs to be reworked.

Go routines that are spawned when the load on a channel increases should
be shut down when the processing is finished, and to correctly do that
Wait groups are required.

Updated Demux to control workers with spawnWorker, thather than go
routines directly.

Context is now being propogated correctly within the demultiplexing
logic.

Need to upgrade how workers are scaled based on monitoring channels.