Added autoscaling and route monitoring logic
Added new functions registerRoute and autoScaler. Both of these generic
functions will allow for a similar implementation of scaling within the
demultiplexer, with the benefit of concurrency hygiene.

registerRoute() will spawn a core worker for each channel, and then
apply the relevant autoscaling parameters for that channel as well.
For instance, the scaling parameters of the friend request channel would
be different to that of encrypted envelopes.