Added demuxRoutes()
Added demuxRoutes() to begin configuring how each payload route should
be processed.

It will contain all the relevant details for how a message type is
processed, and scaled.

Updated processEnvelope() to remove channel processing, and scope it
down to just processing one envelope. ALl processor functions are
subject to this change as the channel processing can be handled by the
demux route bindings.