Added instructions on running Strike without k8s
A user that wants to take a quick look at Strike may not want to go
through the hassle of standing up a k8s cluster.

Using the reinstated make targets to run the server binary locally and an
instance of postgres in docker, there is now instructions on running
strike without k8s.