Started to implement federated presence
Now that there is a skeleton delivery mechanism in place for federated
routing, it's crucial to implement some form of federated presence.

When users connect to a server, that server should announce that user to
other federation members.

Added a new field to FederationOrchestrator to track user presensce on a
given server.

Added basic Lookup/Update functions for users presence on a server.

Added new RPC for Announcing Status among federation. AnnounceStatus{}